Create `pytest` plugin (#454)
Defines a module `pyk.testing` that publishes classes and fixtures used
in the `pyk` test suite for downstream projects.

---------
