+++ src/pyk/testing/__init__.py
@@ -1,0 +1,11 @@
+from ._kompiler import (
+    KCFGExploreTest,
+    KompiledTest,
+    Kompiler,
+    KoreClientTest,
+    KPrintTest,
+    KProveTest,
+    KRunTest,
+    RuntimeTest,
+)
+from ._profiler import Profiler
