+++ src/pyk/testing/plugin.py
@@ -4,10 +4,18 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from .utils import Kompiler
+from ._kompiler import Kompiler
+from ._profiler import Profiler
 
 if TYPE_CHECKING:
+    from pathlib import Path
+
     from pytest import TempPathFactory
+
+
+@pytest.fixture
+def profile(tmp_path: Path) -> Profiler:
+    return Profiler(tmp_path)
 
 
 @pytest.fixture(scope='session')
