+++ src/tests/integration/konvert/test_cell_map.py
@@ -9,8 +9,9 @@ from pyk.konvert import kast_to_kore, kore_to_kast
 from pyk.kore.kompiled import KompiledKore
 from pyk.kore.parser import KoreParser
 from pyk.prelude.kint import INT, intToken
+from pyk.testing import KompiledTest
 
-from ..utils import K_FILES, KompiledTest
+from ..utils import K_FILES
 
 if TYPE_CHECKING:
     from pathlib import Path
