+++ src/tests/integration/kore/test_top_cell_initializer.py
@@ -25,8 +25,7 @@ if TYPE_CHECKING:
     from pathlib import Path
 
     from pyk.kore.syntax import Pattern
-
-    from ..utils import Kompiler
+    from pyk.testing import Kompiler
 
 
 @pytest.fixture(scope='module')
