+++ src/tests/integration/kprove/test_print_prove_rewrite.py
@@ -6,8 +6,9 @@ from pyk.kast.inner import KApply, KRewrite, KVariable
 from pyk.kast.manip import push_down_rewrites
 from pyk.kast.outer import KClaim
 from pyk.prelude.ml import is_top
+from pyk.testing import KProveTest
 
-from ..utils import K_FILES, KProveTest
+from ..utils import K_FILES
 
 if TYPE_CHECKING:
     from pyk.ktool.kprove import KProve
