+++ src/tests/integration/kprove/test_prove_claim_with_lemmas.py
@@ -6,8 +6,9 @@ from pyk.kast.inner import KAtt, KToken
 from pyk.kast.outer import KClaim, KRule
 from pyk.prelude.kbool import BOOL
 from pyk.prelude.ml import is_top
+from pyk.testing import KProveTest
 
-from ..utils import K_FILES, KProveTest
+from ..utils import K_FILES
 
 if TYPE_CHECKING:
     from pyk.ktool.kprove import KProve
