+++ src/tests/profiling/profile_kore_parse.py
@@ -8,7 +8,7 @@ from pyk.kore.parser import KoreParser
 from .utils import TEST_DATA_DIR
 
 if TYPE_CHECKING:
-    from .utils import Profiler
+    from pyk.testing import Profiler
 
 
 def test_kore_parse(profile: Profiler) -> None:
